requirements.txt
colorama
coverage
Django
django-cors-headers
django-nose
djangorestframework
nose
pinocchio
psycopg2-binary
pytz
